Switched to a much simpler method of auto-clicking cards
Instead of the hacky direct API call methods I was using before, this
just checks the state of the DOM to determine if cards are ready to be
clicked, and then activates the click handlers for the appropriate
card elements.

This is way way simpler than what I was trying to do before, and it
updates the state of the UI properly when clicks happen.

It should also be more resilient to changes in their API as it doesn't
require access to their local storage db.